\name{Math}
\alias{Math}
\docType{data}
\title{
Parents' education levels and adolescent mathematics achievement of 76 families in 1986
}
\description{
A dataset containing the mothers' education level, children's mathematical achievement
and other attributes of 76 families.
}
\usage{data("Math")}
\format{
  A data frame with 76 observations on the following 4 variables.
  \describe{
    \item{ME}{mothers' education level}
    \item{HE}{home environment}
    \item{MATH}{children's mathematical achievement}
    \item{READ}{reading recognition ability}
  }
}
\details{
Data used in this example are randomly sampled from the National Longitudinal Survey
of Youth, the 1979 cohort, which were collected in 1986. 
}
\source{
\url{https://www.nlsinfo.org/content/cohorts/nlsy79-children/using-and-understanding-the-data/childyoung-adult-documentation}
}
\references{
Center for Human Resource Research (2006). NLSY79 child & young adult data users guide: a guide to the 1986-2004 child data (Computer software manual). Columbus.
}
\examples{
data(Math)
}
\keyword{datasets}
